Use a non standard Gemfile name
To not interfere with bundler specs.

Currently it doesn't interfere because test stuff is created under
`Dir.tmpdir`, which normally is `/tmp`. However, under some circumstances
`Dir.tmpdir` will be the current directory and in that case the
`Gemfile` will get in the middle and cause tests to fail.

Use a non-standard name instead.