Set `BUNDLE_GEMFILE` while installing dependencies
`Bundler::Installer.install`, even if it receives a
`Bundler::Definition` still depends on the `BUNDLE_GEMFILE` environment
being set globally.

We hadn't noticed this because the parts using it are not essential in
currently there's a `Gemfile` at the root of the repo, so that one gets
(incorrectly) picked up. If we removed that `Gemfile`, things would
crash though.

Removing this global dependency on the environment would be nice, but
for now I'm setting the environment the installer expects.